requirements.txt
azure-storage-file-datalake
azure-identity
boto3
loguru
goodconf